backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
packaging>=21.0
torch>=2.1.1.
torchvision
transformers==4.53.1
timm==0.9.12
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
aiofiles==23.2.1
matplotlib==3.8.2
numpy==1.24.3
opencv-python==4.8.1.78 
einops==0.8.0